requirements.txt
functions-framework==3.0.0
requests==2.28.2
google-cloud-firestore==2.20.0
firebase-admin==6.3.0
google-cloud-secret-manager==2.19.0
flask==2.2.5
gunicorn==20.1.0
pytz
openai==1.20.0
httpx==0.27.0
